Update README.md
This project is a simple Python-based Random Number Generator program that generates random numbers within a user-specified range. The user can specify the starting and ending range, as well as the number of random numbers to generate. The program uses Python's built-in random module, particularly the random.randint() function, to generate the random numbers.